requirement.txt
langchain
langchain_core
langchain_community
langchain_google_genai
python-dotenv